Change partitioning type and add unpartitioning method
Change the type of the partitioning for time from double to long. It
doesn't make much sense to partition on fractions of seconds when we
only store seconds

Add method for unpartitioning to Grid. This gives the location in the
center of the grid.

+++ src/main/scala/swapmob.scala
@@ -11,9 +11,9 @@ object Swapmob {
   type Grids = org.apache.spark.sql.Dataset[GridID]
   type Swaps = org.apache.spark.sql.Dataset[Swap]
 
-  val partitioning: (Double, Double) = (60, 0.001)
+  val partitioning: (Long, Double) = (60, 0.001)
 
-  def partition(data: Measurements, partitioning: (Double, Double)): Grids = {
+  def partition(data: Measurements, partitioning: (Long, Double)): Grids = {
     return data.map(_.partition(partitioning))
   }
